Fix CMAKE_MODULE_PATH issue
When CMAKE_MODULE_PATH was set in directory scope already, the library loader
did not override it. It only wrote a cache variable. Due to CMake's way
it evaluates variables (the cache is used last), the library was not found
because the path was not added. With this commit, the loader writes the
variable in PARENT_SCOPE (directory scope) as well.